requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
click==8.1.3
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
passlib==1.7.4
pydantic==1.10.4
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-multipart==0.0.5
pytz==2022.7.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.2
starlette==0.22.0
tortoise-orm==0.19.3
typing-extensions==4.4.0
uvicorn==0.20.0
